Validation handling with SQS.
